rollout inclue files for mqUtil
git-svn-id: http://syspver1/svn/develop/modules/mqUtil@1479 72f11a4d-5a25-4e8a-b8c8-85b0af744ad3

+++ src/mqbase.c
@@ -421,12 +421,12 @@ int mqGet( MQHCONN _hConn     ,      // connection handle
         *_bufLng = msgLng+1 ;                      //
         _buffer = (PMQVOID) realloc( _buffer ,     //
                                     sizeof(void)*(*_bufLng) );
-        logger( LMQM_INCR_MSG_BUFF, *_bufLng );    //
-        if( _buffer == NULL )                   //
-        {                              //
-          logger( LSTD_MEM_ALLOC_ERROR ) ;      //
-          goto _door ;                    //
-        }                        //
+        logger(LMQM_INCR_MSG_BUFF,(int)*_bufLng ); //
+        if( _buffer == NULL )                      //
+        {                                          //
+          logger( LSTD_MEM_ALLOC_ERROR ) ;         //
+          goto _door ;                             //
+        }                                          //
         break ;                                    //
       }                                            //
                                                    //
